A simple C program to do geometry calculations
Geometry Calculator (Functions)
Write a program that displays the following menu: Geometry Calculator
1. Calculate the Area of a Circle
2. Calculate the Area of a Rectangle 3. Calculate the Area of a Triangle 4. Quit
Enter your choice (1-4):
If the user enters 1,(Void Function) the program should ask for the radius of the circle and then display its area. Use 3.14159 for π. If the user enters 2, (Returning Function taking in parameters; length and width) the program should ask for the length and width of the rectangle, and then display the rectangle’s area. If the user enters 3 (Void Function), the program should ask for the length of the triangle’s base and its height, and then display its area. If the user enters 4, the program should end. Input Validation:
Display an error message if the user enters a number outside the range of 1 through 4 when selecting an
 
item from the menu. Do not accept negative values for the circle’s radius, the rectangle’s length or width, or the triangle’s base or height.